requirements.txt
langchain
langchain-openai
langchain-core
langchain_community
docx2txt
pypdf
langchain_chroma
python-multipart
fastapi
uvicorn
streamlit
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client